chore(app): strip branch comments, move back symbol into icon module
