+++ modules/image-cache-limits/ios/ImageCacheLimitsAppDelegateSubscriber.swift
@@ -12,8 +12,6 @@ import SDWebImage
  * launch as an early pre-JS floor; `index.js` reconfigures the same cache to a
  * RAM-proportional value once the bundle loads. Scale to ~2% of physical RAM,
  * clamped to [96MB, 384MB], so a smaller device caps tighter here too.
- * The Android side lives in the module's ApplicationLifecycleListener: Glide is
- * bounded by default, tiered via setMemoryCategory instead of a byte cap.
  */
 public class ImageCacheLimitsAppDelegateSubscriber: ExpoAppDelegateSubscriber {
   deinit {}
