--- src/constants/backSymbolName.ts
@@ -1,5 +1,0 @@
-export const BACK_SYMBOL_NAME = {
-  ios: 'chevron.left',
-  android: 'arrow_back',
-  web: 'arrow_back',
-} as const;
