+++ src/lib/analytics.ts
@@ -5,10 +5,6 @@ import { logger } from '@app/utils/logger';
 
 const vexoApiKey = process.env.EXPO_PUBLIC_VEXO_API_KEY;
 
-/**
- * No-ops when `analyticsEnabled` is off or no API key is configured, so call
- * sites can fire unconditionally.
- */
 export function trackEvent(
   name: string,
   properties: Record<string, unknown> = {},
