+++ src/screens/Stream/LiveStreamScreen.tsx
@@ -34,9 +34,8 @@ import { Chat } from '@app/components/Chat/Chat';
 import { MEDIA_THUMBNAIL_SIZE } from '@app/components/LiveStreamCard/thumbnailSizes';
 import { StreamPlayer } from '@app/components/StreamPlayer/StreamPlayer';
 import type { StreamPlayerRef } from '@app/components/StreamPlayer/types';
-import { SymbolView } from '@app/components/ui/Icon/Icon';
+import { BACK_SYMBOL_NAME, SymbolView } from '@app/components/ui/Icon/Icon';
 import { Text } from '@app/components/ui/Text/Text';
-import { BACK_SYMBOL_NAME } from '@app/constants/backSymbolName';
 import { useAuthContext } from '@app/context/AuthContext';
 import { useStreamQuery } from '@app/hooks/queries/useStreamQuery';
 import { useUserQuery } from '@app/hooks/queries/useUserQuery';
