+++ src/screens/Stream/components/PlayerBackButton.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next';
 import { router } from 'expo-router';
 
 import { IconButton } from '@app/components/IconButton/IconButton';
-import { BACK_SYMBOL_NAME } from '@app/constants/backSymbolName';
+import { BACK_SYMBOL_NAME } from '@app/components/ui/Icon/Icon';
 import { theme } from '@app/styles/themes';
 
 export function PlayerBackButton() {
